+++ pkg/deployment/reconcile/action_set_member_condition_v2.go
@@ -70,26 +70,44 @@ func (a actionSetMemberConditionV2) Start(ctx context.Context) (bool, error) {
 		as := a.action.Params[setConditionActionV2KeyStatus] == string(core.ConditionTrue)
 
 		if err := a.actionCtx.WithStatusUpdateErr(ctx, func(s *api.DeploymentStatus) (bool, error) {
-			m, _, ok := s.Members.ElementByID(a.action.MemberID)
-			if !ok {
-				a.log.Info().Msg("can not set the condition because the member is gone already")
-				return false, nil
-			}
+			var changed bool
 
-			return m.Conditions.UpdateWithHash(api.ConditionType(aa), as, ar, am, ah), nil
+			s.Members.ForServerGroup(func(group api.ServerGroup, members api.MemberStatusList) error {
+				for i := range members {
+					if members[i].ID == a.action.MemberID {
+						changed = members[i].Conditions.UpdateWithHash(api.ConditionType(aa), as, ar, am, ah)
+						return nil
+					}
+				}
+
+				a.log.Info().Msg("can not set the condition because the member is gone already")
+				return nil
+			}, a.action.Group)
+
+			// If not found then false is returned.
+			return changed, nil
 		}); err != nil {
 			a.log.Warn().Err(err).Msgf("unable to update status")
 			return true, nil
 		}
 	case setConditionActionV2KeyTypeRemove:
 		if err := a.actionCtx.WithStatusUpdateErr(ctx, func(s *api.DeploymentStatus) (bool, error) {
-			m, _, ok := s.Members.ElementByID(a.action.MemberID)
-			if !ok {
-				a.log.Info().Msg("can not set the condition because the member is gone already")
-				return false, nil
-			}
+			var changed bool
 
-			return m.Conditions.Remove(api.ConditionType(aa)), nil
+			s.Members.ForServerGroup(func(group api.ServerGroup, members api.MemberStatusList) error {
+				for i := range members {
+					if members[i].ID == a.action.MemberID {
+						changed = members[i].Conditions.Remove(api.ConditionType(aa))
+						return nil
+					}
+				}
+
+				a.log.Info().Msg("can not remove the condition because the member is gone already")
+				return nil
+			}, a.action.Group)
+
+			// If not found then false is returned.
+			return changed, nil
 		}); err != nil {
 			a.log.Warn().Err(err).Msgf("unable to update status")
 			return true, nil